Ran bootstrap and bootstrap -p scripts.
Added definitions of STDOUT_FILENO and STDERR_FILENO in bsp.h.

Removed the tools subdirectory in c/src/lib/libbsp/patmos/pasim. The tools subdirectory provide additional tools that aid in using the target board, namely a script to automate running the RTEMS Test Suites on the BSP.

For functions boot_card and abort_trap, included the __attribute__ ((used)) attribute in the file headers, and removed the attribute in the actual function declaration.